# backend/docs/swagger.yaml
basePath: /
definitions:
...
  api.problemPayload:
    properties:
      constraints:
        type: string
      difficulty:
        type: string
      maxScore:
        type: integer
      memoryLimitMb:
        type: integer
      published:
        type: boolean
      samples:
        items:
          $ref: '#/definitions/problem.SampleInput'
        type: array
      slug:
        type: string
      statement:
        type: string
      timeLimitMs:
        type: integer
      title:
        type: string
    required:
    - difficulty
    - slug
    - statement
    - title
    type: object
  api.publishPayload:
    properties:
      published:
        type: boolean
    type: object
  api.replaceTestsPayload:
    properties:
      tests:
        items:
          $ref: '#/definitions/api.testCaseDTO'
        type: array
    type: object
...
  api.testCaseDTO:
    properties:
      expected:
        type: string
      input:
        type: string
      ordinal:
        type: integer
      points:
        type: integer
    type: object
...
  problem.Problem:
    properties:
      constraints:
        type: string
      createdAt:
        type: string
      difficulty:
        type: string
      id:
        type: string
      maxScore:
        type: integer
      memoryLimitMb:
        type: integer
      published:
        type: boolean
      slug:
        type: string
      statement:
        type: string
      timeLimitMs:
        type: integer
      title:
        type: string
      updatedAt:
        type: string
    type: object
  problem.ProblemDetail:
    properties:
      constraints:
        type: string
      createdAt:
        type: string
      difficulty:
        type: string
      id:
        type: string
      maxScore:
        type: integer
      memoryLimitMb:
        type: integer
      published:
        type: boolean
      samples:
        items:
          $ref: '#/definitions/problem.Sample'
        type: array
      slug:
        type: string
      statement:
        type: string
      tests:
        items:
          $ref: '#/definitions/problem.TestMetadata'
        type: array
      timeLimitMs:
        type: integer
      title:
        type: string
      updatedAt:
        type: string
    type: object
  problem.Sample:
    properties:
      explanation:
        type: string
      id:
        type: string
      input:
        type: string
      ordinal:
        type: integer
      output:
        type: string
      problemId:
        type: string
    type: object
  problem.SampleInput:
    properties:
      explanation:
        type: string
      input:
        type: string
      ordinal:
        type: integer
      output:
        type: string
    type: object
  problem.TestMetadata:
    properties:
      expectedSha:
        type: string
      id:
        type: string
      inputSha:
        type: string
      ordinal:
        type: integer
      points:
        type: integer
      problemId:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Algorithmic contest platform REST API & judge engine.
  title: MiniAlgothon API
  version: "1.0"
paths:
...
  /api/v1/admin/problems:
    get:
      description: Fetch all problems including draft and published problems.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/problem.Problem'
              type: array
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Admin List All Problems
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Create a new contest problem with metadata and sample test cases.
      parameters:
      - description: Problem definition
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/api.problemPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              $ref: '#/definitions/problem.ProblemDetail'
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Admin Create Problem
      tags:
      - Admin
  /api/v1/admin/problems/{id}:
    delete:
      description: Delete a problem and all associated test cases.
      parameters:
      - description: Problem ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "244":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Admin Delete Problem
      tags:
      - Admin
    get:
      description: Fetch full problem details by ID for editing.
      parameters:
      - description: Problem ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/problem.ProblemDetail'
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Admin Get Problem by ID
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Update existing problem statement, limits, and sample testcases.
      parameters:
      - description: Problem ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated problem payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/api.problemPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/problem.ProblemDetail'
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Admin Update Problem
      tags:
      - Admin
  /api/v1/admin/problems/{id}/publish:
    patch:
      consumes:
      - application/json
      description: Publish or unpublish a problem for contest visibility.
      parameters:
      - description: Problem ID
        in: path
        name: id
        required: true
        type: string
      - description: Publish state
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/api.publishPayload'
      responses:
        "244":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Admin Toggle Published Status
      tags:
      - Admin
  /api/v1/admin/problems/{id}/tests:
    put:
      consumes:
      - application/json
      description: Replace official evaluation test cases for a problem.
      parameters:
      - description: Problem ID
        in: path
        name: id
        required: true
        type: string
      - description: Test cases payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/api.replaceTestsPayload'
      responses:
        "244":
          description: No Content
      security:
      - BearerAuth: []
      summary: Admin Replace Test Cases
      tags:
      - Admin
...
  /api/v1/problems:
    get:
      description: Fetch list of published contest problems for competitors.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/problem.Problem'
              type: array
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List Published Problems
      tags:
      - Problems
  /api/v1/problems/{slug}:
    get:
      description: Fetch detailed problem statement and sample testcases by problem
        slug.
      parameters:
      - description: Problem slug identifier
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/problem.ProblemDetail'
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get Problem by Slug
      tags:
      - Problems
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"